refactor : mountains 파일 구조 정리
- FilterChip, MountainCard를 features/mountains/components/로 분리
- 지역 데이터(REGIONS, DISTRICTS)를 features/mountains/constants/regions.ts로 분리
- CaretDownIcon을 components/icons/로 이동
- mountains.tsx import 경로 업데이트
