+++ features/mountains/components/difficulty-bottom-sheet.tsx
