+++ features/mountains/constants/regions.ts
@@ -1,0 +1,175 @@
+export const REGIONS = [
+  "서울",
+  "경기",
+  "인천",
+  "강원",
+  "대전",
+  "세종",
+  "충남",
+  "충북",
+  "부산",
+  "울산",
+  "경남",
+  "경북",
+  "대구",
+  "광주",
+  "전남",
+  "전북",
+  "제주",
+];
+
+export const DISTRICTS: Record<string, string[]> = {
+  서울: [
+    "서울 전체",
+    "강남구",
+    "강동구",
+    "강북구",
+    "강서구",
+    "관악구",
+    "광진구",
+    "구로구",
+    "금천구",
+    "노원구",
+    "도봉구",
+    "동대문구",
+    "동작구",
+    "마포구",
+    "서대문구",
+    "서초구",
+    "성동구",
+    "성북구",
+    "송파구",
+    "양천구",
+    "영등포구",
+    "용산구",
+    "은평구",
+    "종로구",
+    "중구",
+    "중랑구",
+  ],
+  경기: [
+    "경기 전체",
+    "수원시",
+    "성남시",
+    "고양시",
+    "용인시",
+    "부천시",
+    "안산시",
+    "안양시",
+    "남양주시",
+    "화성시",
+    "과천시",
+  ],
+  인천: [
+    "인천 전체",
+    "중구",
+    "동구",
+    "미추홀구",
+    "연수구",
+    "남동구",
+    "부평구",
+    "계양구",
+    "서구",
+    "강화군",
+    "옹진군",
+  ],
+  강원: [
+    "강원 전체",
+    "춘천시",
+    "원주시",
+    "강릉시",
+    "동해시",
+    "태백시",
+    "속초시",
+    "삼척시",
+  ],
+  대전: ["대전 전체", "동구", "중구", "서구", "유성구", "대덕구"],
+  세종: ["세종 전체"],
+  충남: [
+    "충남 전체",
+    "천안시",
+    "공주시",
+    "보령시",
+    "아산시",
+    "서산시",
+    "논산시",
+    "계룡시",
+    "당진시",
+  ],
+  충북: [
+    "충북 전체",
+    "청주시",
+    "충주시",
+    "제천시",
+    "보은군",
+    "옥천군",
+    "영동군",
+  ],
+  부산: [
+    "부산 전체",
+    "중구",
+    "서구",
+    "동구",
+    "영도구",
+    "부산진구",
+    "동래구",
+    "남구",
+    "북구",
+    "해운대구",
+    "사하구",
+    "금정구",
+    "강서구",
+    "연제구",
+    "수영구",
+    "사상구",
+    "기장군",
+  ],
+  울산: ["울산 전체", "중구", "남구", "동구", "북구", "울주군"],
+  경남: [
+    "경남 전체",
+    "창원시",
+    "진주시",
+    "통영시",
+    "사천시",
+    "김해시",
+    "밀양시",
+    "거제시",
+    "양산시",
+  ],
+  경북: [
+    "경북 전체",
+    "포항시",
+    "경주시",
+    "김천시",
+    "안동시",
+    "구미시",
+    "영주시",
+    "영천시",
+    "상주시",
+    "문경시",
+    "경산시",
+  ],
+  대구: [
+    "대구 전체",
+    "중구",
+    "동구",
+    "서구",
+    "남구",
+    "북구",
+    "수성구",
+    "달서구",
+    "달성군",
+  ],
+  광주: ["광주 전체", "동구", "서구", "남구", "북구", "광산구"],
+  전남: ["전남 전체", "목포시", "여수시", "순천시", "나주시", "광양시"],
+  전북: [
+    "전북 전체",
+    "전주시",
+    "군산시",
+    "익산시",
+    "정읍시",
+    "남원시",
+    "김제시",
+  ],
+  제주: ["제주 전체", "제주시", "서귀포시"],
+};
